Factored out "Panels" Relates #25
